refactoring data_model classes
