+++ setup.py
@@ -1,7 +1,6 @@
+from setuptools import setup
 
-import setuptools
-
-setuptools.setup(
+setup(
     name="dbignite",
     version="0.0.1",
     author="Nathan Buesgens, Amir Kermany, Rachel Sim",
@@ -13,16 +12,6 @@ setuptools.setup(
         "License :: databricks :: OSS",
         "Operating System :: OS Independent",
     ],
-    package_dir={"": "dbignite"},
-    packages=setuptools.find_packages(where="dbignite"),
-    python_requires=">=3.6",
+    packages=['dbignite'],
+    py_modules=['dbignite.data_model']
 )
-
-# from setuptools import setup
-
-# setup(
-#     name='dbignite',
-#     version='',
-#     packages=['dbignite'],
-#     py_modules=['dbignite.data_model']
-# )
